# Add tests for measuring effect of lsn caching