See #1949