@hlinnaka 2022-09-02
please fix the typos in commit message before poshing ("measurif" -> "measuring", "effet" -> effect"