@hlinnaka 2022-09-05
In the future, please fix the commit message before pushing. The commit message for this was now "Add test for last written lsn cache", but that was very wrong. The test was added in the earlier commit already and this actually included the code changes.

Results from the new performance test from CI:

[Before](https://github.com/neondatabase/neon/runs/8172060005):
```
test_measure_read_latency_heavy_write_workload[neon-1].read_latency_max: 1.469 s
test_measure_read_latency_heavy_write_workload[neon-1].read_latency_avg: 0.454 s
test_measure_read_latency_heavy_write_workload[neon-1].read_latency_stdev: 0.478
```
[After](https://github.com/neondatabase/neon/runs/8178930935):
```
test_measure_read_latency_heavy_write_workload[neon-1].read_latency_max: 1.408 s
test_measure_read_latency_heavy_write_workload[neon-1].read_latency_avg: 0.265 s
test_measure_read_latency_heavy_write_workload[neon-1].read_latency_stdev: 0.439 s
```
That's a bit disappointing, the effect that was reported initially at https://github.com/neondatabase/neon/pull/1949 were much bigger. Did we make some other changes that mostly fixed the problem already?