@knizhnik 2022-09-05
> Did we make some other changes that mostly fixed the problem already?

I do not know such optimization.
Only changing `max_replication_write_lag` can significantly affect results.
Right now for tests  `max_replication_write_lag=15MB` is used and looks like this lag is small enough to keep response time small enough even without LSN cache.
The table below contains dependency of `read_latency_max` on `max_replication_write_lsg`:
| max_replication_write_lsg | with cache | without cache |
| ----------------------------------- | ------------- | ------------------- |
| 10 | 1.478  | 4.993 |
| 20 | 1.676  | 5.257 |
| 30 | 2.204 | 16.591 |
| 40 | 2.472  | 20.421 | 
| 50 | 2.938  | 22.3162 |