@knizhnik 2022-09-05
Unfortunately I still got one "key not found error" in one CI run:
```
test_runner/regress/test_parallel_copy.py:55: in test_parallel_copy
    asyncio.run(parallel_load_same_table(pg, n_parallel))
/home/nonroot/.pyenv/versions/3.9.2/lib/python3.9/asyncio/runners.py:44: in run
    return loop.run_until_complete(main)
/home/nonroot/.pyenv/versions/3.9.2/lib/python3.9/asyncio/base_events.py:642: in run_until_complete
    return future.result()
test_runner/regress/test_parallel_copy.py:39: in parallel_load_same_table
    await asyncio.gather(*workers)
test_runner/regress/test_parallel_copy.py:29: in copy_test_data_to_table
    await pg_conn.copy_to_table(table_name, source=copy_input)
/github/home/.cache/pypoetry/virtualenvs/zenith-_pxWMzVK-py3.9/lib/python3.9/site-packages/asyncpg/connection.py:877: in copy_to_table
    return await self._copy_in(copy_stmt, source, timeout)
/github/home/.cache/pypoetry/virtualenvs/zenith-_pxWMzVK-py3.9/lib/python3.9/site-packages/asyncpg/connection.py:1081: in _copy_in
    return await self._protocol.copy_in(
asyncpg/protocol/protocol.pyx:529: in copy_in
    ???
E   asyncpg.exceptions.PostgresIOError: could not read block 2 in rel 1663/12972/16384.0 from page server at lsn 0/016960E8
E   DETAIL:  page server returned error: could not find data for key 000000067F000032AC0000400000FFFFFFFF at LSN 0/1696070, for request at LSN 0/16960E8
```
Not sure if it is related with last written LSN cache but it is ost probable candidate.
Any attempts to reproduce it locally still failed.